feat(profiling): Unify aggregate flamegraph sidepanels
Unifies the aggregate flamegraph sidepanels to always show the currently
selection function and a list of profiles with that function.
